Day 02
I learned: use the Rank operator `"` to allow you to apply a verb to each row in an array.
In today's code I use `verb "1 array`.
I also use `]` to break two numbers that would otherwise be treated as an array. Im defining the array inline as `_4 [\ input`.  I need the `]` to separate `verb "1` from `_4` so it is interpreted as (verb "1) (_4 [\ input) and not (verb "(1 _4)) ... .