03 part one done
Roadblock : working code in editor didn't work when run as a script.  If
a user-defined verb is to be called from within another definition, it
seems I need to use a global rather than local define.

I've pulled the code to match letters from the J phrasebook. I'll need
to understand properly how that works in order to extend it for part
two.